requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
# psycopg2-binary==2.9.9  # Optional: Only needed for PostgreSQL
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
jinja2==3.1.2
python-jose[cryptography]==3.3.0
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
boto3==1.34.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.2
faker==20.1.0
python-dotenv==1.0.0
httpx==0.25.2
email-validator==2.1.0
psycopg2-binary
pymysql==1.1.1
